Создать таксопарк

+++ Modul2HW6/Modul2HW6/models/FuelConsumption.cs
@@ -1,5 +1,5 @@
 ﻿public class FuelConsumption
 {
-    public int Value { get; set; }
+    public double Value { get; set; }
     public FuelConsumptionUnit FuelConsumptionUnit { get; set; }
 }
